server/app/requirements.txt
flask
flask-mysql
pycryptodome
cryptography